Code cleanup and dag generation correct logging
