+++ cmd/karlsenminer/main.go
@@ -29,7 +29,7 @@ func main() {
 
 	// Show version at startup.
 	log.Infof("Version %s", version.Version())
-	log.Infof("Using KarlsenHash V2")
+	log.Infof("Using KarlsenHashV2 impl: %s", hashingAlgoVersion)
 
 	// Enable http profiling server if requested.
 	if cfg.Profile != "" {
