+++ domain/consensus/utils/pow/log.go
@@ -5,7 +5,5 @@ import (
 	"github.com/karlsen-network/karlsend/util/panics"
 )
 
-// var log = logger.RegisterSubSystem("POW")
-var backendLog = logger.NewBackend()
-var log = backendLog.Logger("POW")
+var log = logger.RegisterSubSystem("POW")
 var spawn = panics.GoroutineWrapperFunc(log)
